Added "production outputs."
This really isn't anything special, just the parser can now write the
productions taken to a file.

We need to figure out how to mke these outputs coincide with the line
seen by the scanner.

There may be other states to add, i.e. for the function pro- and epilogs.

We should figure out how to keep track of the type we are most recently
working with, etc.