Added some useful stuff and some unfinished stuff.
Added the PYTHON_PATH modification to be able to run jstcc.py from
anywhere.
Added a simple logger class that might be helpful to us.
Started adding symbol table logic into the parser. (This required at
least a dummy SymbolBuilder, feel free to replace). Not finished yet;
asked Fred for help on how to do it.