Made A LOT of changes.
Added what I hope is the interface to the Symbol class for building up a symbol as it moves up the parse tree.

Added what I hope the interface will be for the SymbolTable for things like looking up types. TODO: Add tracking of typedef types to the Scope class (maybe this is part of Symbol, maybe not?)

Made various minor fixes to the lexer.

Modified the lexer to take out unused token types (i.e. RANGE). I think somewhere along the line we should push error handling into the parser as much as possible to unify error management.

Made the "plain main" test of the scanner/lexer test what tokens are found. We should update all tests accordingly.

Added just a little to the symbol_table package's tests.