Calling it a night, back at it in the morning.
Fixed some of the grammar for struct declarations (I think autocomplete screwed things up).

Added some test cases around tokenizing comments to the lexer. Fixed eating of single-line comments.

Working on parsing typedefs - still failing on the lookup when we use them. Need to implement the type storing part of the symbol table in order for these to work.