Refactored the Parser and Lexer to follow PLY's documented format.
We had some weird mix of an encapsulated parser and lexer. The PLY documentation says that we can create separate classes and pass them in as modules, but we weren't quite doing that. Both the lexer and the parser had functionality that they shouldn't have been implementing (possibly to fix errors in the console?). I fixed that and updated as much of the test files and source code as I could.

The new "system" has a JSTLexer and a JSTParser which are the modules for the PLY Lexer and Parser, respectively. CompilerState now handles the main call to parse and teardown.

Aside from the Lexer tests, all other tests are passing as they previously did. I will work with Janelle to get the Lexer tests fixed because I don't understand how they work.