Refactored the declaration and initialization of variables and arrays.
I also added warnings for variables (redeclaration and shadowing) and warnings for arrays (missing dimensions, non-integral dimensions).

I broke functions a little bit by making one of the productions (direct_declarator) return a dict containing the identifier instead of the string identifier itself. I'll be working on this again tomorrow, so it should be fixed eventually.