Finished non-parametrized function declaration and implementation.
Redeclaration of functions is checked in
  declaration : declaration_specifiers init_declarator_list SEMI
Reimplementation of function is checked in
  function_definition : declaration_specifiers declarator enter_function_scope compound_statement leave_scope

FunctionDefinition no longer has a nested FunctionDeclaration node. FunctionDeclaration is only created when a function is declared.

Added an initial enter_scope to the program and a leave_scope to function_definition because scopes were not being left at the end of a function declaration. (Compound statement only left its own scope).