Went through all parser tests to verify everything was being parsed correctly.
Many of the tests were failing, but I was able to get most of them fixed.

They're mainly related to function calls and matching types of passed parameters to the function signature.

We also need to start looking at verifying types for assignment operations. The same production(s) also get used in calling functions.

I started working on a refactor to make the Constant AST node be a special case of VariableSymbol. I had a couple observations and thoughts.
(1) Constants are AST nodes, but VariableSymbols are not. All classes under Symbol can possibly become AST nodes.
(2) VariableSymbol needs to hold a value (for a constant, initialization, or current value).
(3) By implementing (2), we can wrap constants in VariableSymbols (but with what identifier?), perform the same type checking, and eliminate some checks for Constant vs. Symbol.