Various cleaning. See slack for the long descriptions.
Moved unused AST node classes to another file for holding.

Cleaned up the utils stuff to make it a little more wieldy.

Analyzed some of the stuff for type checking and made a start on improving it. The main change being that any AST node that could be the root of an expression must implement get_resulting_type to return the type that would be typechecked.

There's probably more, but I can't for the life of me remember what else. More tests are passing, so hopefully I'm on the right track.