Modified get_promoted_type
By original design, it seemed to treat the sign difference as a cast. I just modified it to be a downcast. I also modified the method signature to be left and right types. This method can (and probably) will replace can_assign() from the now-removed AssignmentUtils.

My rationale for returning just the result and not a (type, result) tuple is that the type of the left is not going to be modified since it is a lvalue, but the right side can since it's a rvalue. Therefore, all cast results indicate what should happen to the type on the right.

An example:
int a = float b  // b has to be down cast to int
float a = int b  // b has to be up cast to float