Implemented the 'RegisterUseTable'.
This data structure manages mapping of pseudo-registers (aka temporaries) to real/physical registers (like the MIPS $t0 register). The data structure also generates the necessary code to facilitate register spilling.

Usage summary: The RegisterUseTable should be used _often_. That is calls to acquire() should be used anytime a psuedo-register is about to be used. Calls to acquire() return a dictionary containing the name of the physical register that has been made available and the assembler code that should be used to ensure that the register is available for use. Once one can be sure that a pseudo-register is not in use (for us, that is likely only at the beginning and end of a function call) a call to release() or release_all() should be made to help things along in the register management process.

Only basic testing has been implement it. More rigorous testing should be implemented, but it would be good to commit earlier and have a 'checkpoint.'