Added the use of 3AC argument types: Register, Immediate, and Address.
These classes should help us out by carrying information about the argument through the program, instead of having to extract it out of a string or something via regular expressions. This is most apparent in the case of Address.

I have only used these sparingly to avoid work and breaking things where possible, but I think these will help a lot, especially in places where we need to ensure the type of an argument. I used them in the LOAD and STORE instructions to guarantee that an Address is passed in. Use in other places is probably merited.